seperated payments to enrollments app
