+++ core/api/admin.py
@@ -2,13 +2,12 @@ from django.contrib import admin
 
 # Register your models here.
 
-from .models import Contributions, ContributionTags, contributionVideos, ContributionNotes, Enrollment, ContributionsComments, ContributionRatings,University,Department,MajorSubject
+from .models import Contributions, ContributionTags, contributionVideos, ContributionNotes, ContributionsComments, ContributionRatings,University,Department,MajorSubject
 
 admin.site.register(Contributions)
 admin.site.register(ContributionTags)
 admin.site.register(contributionVideos)
 admin.site.register(ContributionNotes)
-admin.site.register(Enrollment)
 admin.site.register(ContributionsComments)
 admin.site.register(University)
 admin.site.register(Department)
