+++ core/core/settings.py
@@ -51,6 +51,7 @@ INSTALLED_APPS = [
     'api',
     'rest_framework_simplejwt',
     'rest_framework_simplejwt.token_blacklist',
+    'enrollments'
 ]
 
 MIDDLEWARE = [
