+++ core/enrollments/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+# Register your models here.
+from .models import Enrollment
+
+
+admin.site.register(Enrollment)
